+++ composeApp/src/commonMain/kotlin/com/foodapp/foodapp/presentation/userScreen/mainScreen/screens/bookingScreen/BookingScreenViewModel.kt
@@ -47,7 +47,7 @@ class BookingScreenViewModel(private val screenSize: PlatformConfiguration,
 
     }
 
-    fun getScreenSize():Pair<Int,Int>{
+    fun getScreenSize():Pair<Float,Float>{
         return Pair(screenSize.screenWidth(),screenSize.screenHeight())
     }
 }
