+++ composeApp/src/commonMain/kotlin/com/foodapp/foodapp/presentation/userScreen/mainScreen/screens/cartScreen/CartScreenViewModel.kt
@@ -130,7 +130,7 @@ class CartScreenViewModel(
         }
     }
 
-    fun getScreenSize(): Pair<Int, Int> {
+    fun getScreenSize(): Pair<Float, Float> {
         return Pair(screenSize.screenWidth(), screenSize.screenHeight())
     }
 }
