+++ composeApp/src/commonMain/kotlin/com/foodapp/foodapp/presentation/userScreen/mainScreen/screens/restaurantScreen/ViewRestaurantScreenViewModel.kt
@@ -20,7 +20,7 @@ class ViewRestaurantScreenViewModel(private val screenSize:PlatformConfiguration
         }
     }
 
-    fun getScreenSize():Pair<Int,Int>{
+    fun getScreenSize():Pair<Float,Float>{
         return Pair(screenSize.screenWidth(),screenSize.screenHeight())
     }
